KVM: Fix RMW mmio handling
Commit 9bf671a47ed6af3164524a31dbef9360f1b66fb5 optimized the mmio
read path by returning to the emulator directly after an mmio read request.
But we may also need to return back to userspace in case the instruction
was a read-modify-write instruction, which means we need to issue a write
after completion of the read instead of returning to the guest.

Signed-off-by: Avi Kivity <avi@qumranet.com>