KVM: Move interrupt injection out of interrupt disabled section
Instead of injecting interrupts while we're in the critical section during
the guest switch, inject them earlier.  In case we had an irq raised between
the injection point and the switch, we check a bit during guest switch and
if needed we go back and redo the injection.

This improves system latency, and allows sleeping during injection (which
is needed when injecting real-mode interrupts on Intel).

Signed-off-by: Avi Kivity <avi@qumranet.com>