KVM: VMX: Add mechanism to detect VMentry failure
Adds mechanism to detect VMentry failure. When VMentry
failure is detected we're trying to emulate the instruction
that causes this failure.  If emulation failed we modify
the SS selector to pass VMX sanity tests and retry to
execute the instruction (old behaviour).  The good solution
is to emulate the instruction that causes the failure.

Signed-off-by: Guillaume Thouvenin <guillaume.thouvenin@ext.bull.net>
Signed-off-by: Laurent Vivier <laurent.vivier@bull.net>
Signed-off-by: Avi Kivity <avi@qumranet.com>