KVM: VMX: Enable NMI window reporting unconditionally if deferring nmi injection
If we are deferring an NMI injection due to an interrupt being reinjected,
enable the NMI window unconditionally, so we don't have excess latency for
the NMI injection.  Previously we only enabled the NMI window if it was
previously closed, which is a rare and unintersting situation.

Noted by Jan Kiszka.

Signed-off-by: Avi Kivity <avi@redhat.com>