KVM: SVM: Fix bogus sysenter MSRs in userspace
When we read and write the sysenter MSRs we copy them to private variables
which are never seen by the guest.  This causes live migration to fail
when the guest performs a sysenter or sysexit for the first time on the
destination.

Fix by reading and writing the values directly to the vmcb.

Signed-off-by: Avi Kivity <avi@redhat.com>