KVM: SVM: Fix sysenter migration issue on AMD CPUs
To enable cross-vendor migration we use VMCB external variables to hold
the full 64bit value of the SYSENTER MSRs, which get truncated to 32bit
on AMD hardware. Since we didn't intercept these MSRs, these
variables were only used in the emulation case, but were _always_ used
for migration purposes. This worked fine for cross-vendor migration in
compat mode, but did not work in pure legacy mode.

To fix this we always intercept the SYSENTER MSRs and store the values
both in the VMCB and the external variables. This works for all cases.

Signed-off-by: Andre Przywara <andre.przywara@amd.com>
Signed-off-by: Avi Kivity <avi@redhat.com>