KVM: PPC: E500 compile fix
While trying to compile an E500 vmlinux, I stumbled across a compilation bug
that was obviously there before I touched any of the code. A trace point
doesn't get the correct arguments.

Since that shouldn't be any critical to the functionality of the code, my quick
workaround is to #if 0 it out.

Signed-off-by: Alexander Graf <agraf@suse.de>
CC: Liu Yu <yu.liu@freescale.com>
Signed-off-by: Avi Kivity <avi@redhat.com>