ARM: Allow FIQs support with the domains removal patch
Once domains have been removed, the vectors page is mapped as user RO,
kernel RO. However, the FIQ support relies on being able to copy code to
this page at run-time. This commit saves the address of the vectors page
in a dedicated variable that will be accessed directly by the
set_fiq_handler() and early_trap_init().

Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>