ARM: Do not create temporary mapping when populating the vectors page
With the domains removal patch, the vectors page is no longer writable.
The original implementation was creating a temporary writable mapping of
the vectors page. Since FIQs support requires a permanently writable
vectors page (at a different address), this commit changes the
early_trap_init() implementation to access the vectors page alias
directly. It is assumed that there are no cache aliasing issues on such
platforms.

Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>