ARM: vexpress: add support for ct-ca15x4 CLCD controller
This patch adds support for the CLCD controller present on
the ct-ca15x4 core tile.

*** NOTE ***
Currently the model doesn't provide *anything* in the core-tile
space so we can't probe for a device without causing an external
abort. This patch also includes a workaround for a model bug
when accessing the DVI mux on the motherboard.

Signed-off-by: Will Deacon <will.deacon@arm.com>