KVM: (arm) Don't register MMIO regions
It's much simpler, and the way it's done in other architecture, if we
instead of explicitly mapping MMIO regions to KVM simply forward all
accesses to unknown memory to QEMU and let it emulate whatever it needs.

This has the improved benefit of simplifying the abort handler code
somewhat.

Note: This commit should be used by the upstream version of QEMU, which
has been modified to support KVM for ARM on the armkvm-upstream branch,
in the android-virt mirror of QEMU.