KVM: (arm) Fix the abort handler for prefetch aborts on v5
ARMv5 doesn't update the FSR on prefetch aborts, but
the emulated QEMU architecture updates IFSR instead. However,
to better support the architectures, there should be a method
to determine the reason for the fault even if there is no
FSR value available - will be interesting to see what's the
solution for ARMv6.