KVM: (arm) Asembly cleanup, cache/tlb fixes and debugging
Changed the header in the shared page to use the size of the struct as
it is defined in C and using macros to calculate offsets instead of
referring to labels, which must be synchronized with the C struct in
hand.

Also used alternative cache cleaning and TLB invalidation functions from
boot code instead of from context switch code, and differences should be
investigated before put back into main branches.

Finally a lot of debugging functionality, which should also be removed
from this commit.