KVM: (arm) cleanup some handle_swi code
If the instruction following a guest SWI (hypercall) was in the same
page as the SWI instruction, the exception vector copied it directly to
the shared page when the guest page tables were in use, thereby avoiding
aliases and better utilize TLBs and caches. If this was the case, the
handle_swi function should just use this value instead of copying it
from user space.