KVM: (arm) Getting rid of memcopies...
Thie patch was originally crafted by Jeremy Andrus, but got a littel
dated as it was written for the original ARMv5 environment.

Anyway, the general idea is to simply to all the vcpu registers reguired
during the world-switches directly and permanently on the shared page
and thereby avoiding copying approx. 100 bytes at each entry and exit to
the guest.

Since accesing VCPU registers is the common operation in all of the
code, this patch is quite intrusive, which on the other hand was a good
excuse to do a little cleanup. More of that will follow.