KVM: (arm) Use SRS and SVC r13 in exception return
We used to store temporary values directly on the exception vector page,
but this adds an extra page in the cache, which is also already in the
I-cache. Additionally we save a few instructions by using the SRS
instructions.

WARNING: There is an undiscovered bug in this commit causing continous
prefetch aborts on 0x0 to occur.