KVM: (arm) Trace and performance counting
Refactored a lot of the debugging/tracing code from arm.c to separate
trace.c file in arch/arm/kvm/...

Also introduced a simple array of events which are counted and printed
in a sorted maner to be able to tell where the time is spent. This could
be easily expanded to read a cycle counter and give more accurate
metrics.

Additionally, much of this should actually be done using the kernel perf
tool and tracepoints. However, the way that the rest of KVM does it
relies on newer kernel features than 2.6.29, and serial console on the
development devices does not yet work with anything newer, so we do this
quick and dirty for now. Before upstream, this should be improved.