KVM: (arm) Update APs instead of flushing shadows
When the VCPU changes between privileged and user mode, the access
permission bits must change on the shadow page tables, since they are
adapted for running always in user mode. Until now, we have simply
invalidated the entire shadow page table when this happened causing
quite a performance degredation when the guest started running in user
space and issuing system calls.

This commit introduces a new routine, which scans the shadow page table
and for each entry find the corresponding guest entry and re-calculates
the correct page table entries.

Since this reads guest page tables several times, it's too expensive to
clean caches on each run. Instead, we move the cache cleaning when
reading shadow page tables to the callers, which requires only a single
cleaning operation for the entire AP switch.