KVM: (arm) Handle conditionals in emulate priv. dp.
When emulating privileged data processing instructions, the instructions
may be configured to execute conditionally, and for this to work
correctly we need the CPSR cond bits to be that of the guest state when
emulating the instruction on the host side. This is in fact not used by
Linux guest versions tested so far, but it was a 'theoretical bug'.