KVM: (arm) Fix vector access bug
Some times the host would catch an error in which the guest could not
access its own vector page even in privileged mode. This would typically
happen after switching the vector location for KVM a couple of times.

Turned out to be very simple, needed to clean caches after unmapping
addresses in the shadow page tables.

This commit is related to the 'KVM: (arm) Debug vector access and guest
bug' commit in that it fixes what the previous tries to debug and also
keeps some of the introduced functionality around although renaming it a
little bit to be more clear for future use. It will probaby go away
eventually though.