KVM: (arm) Better cache cleaning handling
Factor out some functionality to a header file that considers the
current implementation of cache behavior and delegates the
responsibility. Also start using the Linux macros to flush the pmd
entries when modifying shadow page tables to ensure correctness at that
end and clean up a few other places regarding cache maintenance, which
were commented out and so on.