KVM: (arm) Emulate DP instructions on v7 (and v6 improv.)
The Data processing instruction emulation is done using self-modifying
code and the move to ARMv7 requires special handling of this, which
differs from that of ARMv6. However, in the process, the code for ARMv6
could be written a little nicer, and that change is also a part of this
commit.