ARM: architected timers: move local timer support to percpu_timer.c
To introduce the A15 architected timers without causing too
much disruption on the rest of the kernel, introduce percpu_timer.c
to act as a registration interface that the platform can call at
runtime.

Timer functions are moved out of smp.c (as the A15 can use the
local timer in UP configuration) and  smp_twd.c now returns
a set of operations that percpu_timer.c can call.

Acked-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>