kvm tools: Indicate the end of a KVM session
Right now when a KVM session ends with an incomplete line printk'd
without a newline, kvm exits in an ugly way:

  [    4.638016] Kernel panic - not syncing: VFS: Unable to mount root fs on unknown-block(0,0)
  [    4.641668] Rebooting in 1 seconds..spirit:~/tip/tools/kvm>
  spirit:~/tip/tools/kvm>

See the shell prompt intermixed with the last line of kernel output.

There's also no indication to the user that everything is fine on the
kvm tool side.

This is somewhat of a beauty wart and also a bit confusing to users.

To clarify and clean up things, add a final printout on normal exits:

  [    4.654814] Please append a correct "root=" boot option; here are the available partitions:
  [    4.584390] Kernel panic - not syncing: VFS: Unable to mount root fs on unknown-block(0,0)
  [    4.662243] Rebooting in 1 seconds..
    # KVM session ended normally.
  spirit:~/tip/tools/kvm>

This delimits the kernel output properly and gives feedback to the user
that (despite the scary kernel messages) all fine from the KVM POV.

Signed-off-by: Ingo Molnar <mingo@elte.hu>
Signed-off-by: Pekka Enberg <penberg@kernel.org>