ARM: reset: add reset functionality for jumping to a physical address
Tools such as kexec and CPU hotplug require a way to reset the processor
and branch to some code in physical space. This requires various bits of
jiggery pokery with the caches and MMU which, when it goes wrong, tends
to lock up the system.

This patch implements a new function, arm_machine_reset, for
consolidating this code in one place where it can be used by multiple
subsystems. A static array is used for the stack during reset so that
the identity mapping doesn't get in the way.

Signed-off-by: Will Deacon <will.deacon@arm.com>